+++ sdk/core/Azure.Core.Experimental/src/DynamicJson.cs
@@ -24,10 +24,12 @@ namespace Azure.Core.Dynamic
         private static readonly MethodInfo SetViaIndexerMethod = typeof(DynamicJson).GetMethod(nameof(SetViaIndexer), BindingFlags.NonPublic | BindingFlags.Instance)!;
 
         private MutableJsonElement _element;
+        private DynamicJsonOptions _options;
 
-        internal DynamicJson(MutableJsonElement element)
+        internal DynamicJson(MutableJsonElement element, DynamicJsonOptions options = default)
         {
             _element = element;
+            _options = options;
         }
 
         internal override void WriteTo(Stream stream)
@@ -39,12 +41,44 @@ namespace Azure.Core.Dynamic
 
         private object? GetProperty(string name)
         {
+            Argument.AssertNotNullOrEmpty(name, nameof(name));
+
             if (_element.TryGetProperty(name, out MutableJsonElement element))
             {
                 return new DynamicJson(element);
             }
 
+            if (PascalCaseGetters() && char.IsUpper(name[0]))
+            {
+                if (_element.TryGetProperty(GetAsCamelCase(name), out element))
+                {
+                    return new DynamicJson(element);
+                }
+            }
+
             return null;
+        }
+
+        private bool PascalCaseGetters()
+        {
+            return
+                _options.PropertyNameCasing == DynamicJsonNameMapping.PascalCaseGetters ||
+                _options.PropertyNameCasing == DynamicJsonNameMapping.PascalCaseGettersCamelCaseSetters;
+        }
+
+        private bool CamelCaseSetters()
+        {
+            return _options.PropertyNameCasing == DynamicJsonNameMapping.PascalCaseGettersCamelCaseSetters;
+        }
+
+        private static string GetAsCamelCase(string value)
+        {
+            if (value.Length < 2)
+            {
+                return value.ToLowerInvariant();
+            }
+
+            return $"{char.ToLowerInvariant(value[0])}{value.Substring(1)}";
         }
 
         private object? GetViaIndexer(object index)
@@ -62,11 +96,47 @@ namespace Azure.Core.Dynamic
 
         private IEnumerable GetEnumerable()
         {
-            return new ArrayEnumerator(_element.EnumerateArray());
+            return _element.ValueKind switch
+            {
+                JsonValueKind.Array => new ArrayEnumerator(_element.EnumerateArray()),
+                JsonValueKind.Object => new ObjectEnumerator(_element.EnumerateObject()),
+                _ => throw new InvalidOperationException($"Unable to enumerate JSON element."),
+            };
         }
 
         private object? SetProperty(string name, object value)
         {
+            Argument.AssertNotNullOrEmpty(name, nameof(name));
+
+            if (_options.PropertyNameCasing == DynamicJsonNameMapping.None)
+            {
+                _element = _element.SetProperty(name, value);
+                return null;
+            }
+
+            if (!char.IsUpper(name[0]))
+            {
+                // Lookup name is camelCase, so set unchanged.
+                _element = _element.SetProperty(name, value);
+                return null;
+            }
+
+            // Other mappings have PascalCase getters, and lookup name is PascalCase.
+            // So, if it exists in either form, we'll set it in that form.
+            if (_element.TryGetProperty(name, out MutableJsonElement element))
+            {
+                element.Set(value);
+                return null;
+            }
+
+            if (_element.TryGetProperty(GetAsCamelCase(name), out element))
+            {
+                element.Set(value);
+                return null;
+            }
+
+            // It's a new property, so set according to the mapping.
+            name = CamelCaseSetters() ? GetAsCamelCase(name) : name;
             _element = _element.SetProperty(name, value);
 
             // Binding machinery expects the call site signature to return an object