+++ sdk/storage/Azure.Storage.Files.Shares/src/ShareClientOptions.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using Azure.Core;
+using Azure.Storage.Files.Shares.Models;
 
 namespace Azure.Storage.Files.Shares
 {
@@ -102,7 +103,12 @@ namespace Azure.Storage.Files.Shares
             /// <summary>
             /// The 2021-12-02 service version.
             /// </summary>
-            V2021_12_02 = 15
+            V2021_12_02 = 15,
+
+            /// <summary>
+            /// The 2022-11-02 service version.
+            /// </summary>
+            V2022_11_02 = 16
 #pragma warning restore CA1707 // Identifiers should not contain underscores
         }
 
@@ -118,6 +124,34 @@ namespace Azure.Storage.Files.Shares
         /// Transfer validation options to be applied to file transfers from this client.
         /// </summary>
         public TransferValidationOptions TransferValidation { get; } = new();
+
+        /// <summary>
+        /// If set to true, trailing dot (.) will be allowed to suffex directory and file names.
+        /// If false, the trailing dot will be trimmed.
+        /// Supported by x-ms-version 2022-11-02 and above.
+        /// </summary>
+        public bool? AllowTrailingDot { get; set; }
+
+        /// <summary>
+        /// If set to true, trailing dot (.) will be allowed to source file names.
+        /// If false, the trailing dot will be trimmed.
+        /// Supported by x-ms-version 2022-11-02 and above.
+        /// Applicable to <see cref="ShareFileClient.Rename(string, Models.ShareFileRenameOptions, System.Threading.CancellationToken)"/>,
+        /// <see cref="ShareFileClient.RenameAsync(string, Models.ShareFileRenameOptions, System.Threading.CancellationToken)"/>,
+        /// <see cref="ShareFileClient.UploadRangeFromUri(System.Uri, HttpRange, HttpRange, Models.ShareFileUploadRangeFromUriOptions, System.Threading.CancellationToken)"/>,
+        /// <see cref="ShareFileClient.UploadRangeFromUriAsync(System.Uri, HttpRange, HttpRange, Models.ShareFileUploadRangeFromUriOptions, System.Threading.CancellationToken)"/>,
+        /// <see cref="ShareFileClient.StartCopy(System.Uri, Models.ShareFileCopyOptions, System.Threading.CancellationToken)"/>,
+        /// <see cref="ShareFileClient.StartCopyAsync(System.Uri, Models.ShareFileCopyOptions, System.Threading.CancellationToken)"/>,
+        /// <see cref="ShareDirectoryClient.Rename(string, Models.ShareFileRenameOptions, System.Threading.CancellationToken)"/>,
+        /// and <see cref="ShareDirectoryClient.RenameAsync(string, Models.ShareFileRenameOptions, System.Threading.CancellationToken)"/>.
+        /// </summary>
+        public bool? AllowSourceTrailingDot { get; set; }
+
+        /// <summary>
+        /// Share Token Intent.  For use with token authentication.  Used to indicate the intent of the request.
+        /// This is currently required when using token authentication.
+        /// </summary>
+        public ShareTokenIntent? ShareTokenIntent { get; set; }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="ShareClientOptions"/>