+++ sdk/botservice/Azure.ResourceManager.BotService/CHANGELOG.md
@@ -1,14 +1,21 @@
 # Release History
 
-## 1.0.0-beta.2 (Unreleased)
+## 1.0.0 (2023-03-07)
+
+This release is the first stable release of the BotService Management client library.
 
 ### Features Added
 
+- Added Update methods in resource classes.
+
 ### Breaking Changes
 
-### Bugs Fixed
+Polishing model names and resource names to make them more descriptive since last public beta release.
 
 ### Other Changes
+
+- Upgraded dependent `Azure.Core` to `1.28.0`.
+- Upgraded dependent `Azure.ResourceManager` to `1.4.0`.
 
 ## 1.0.0-beta.1 (2022-09-15)